exercises complete

+++ src/HelloWorld.java
@@ -24,8 +24,41 @@ public class HelloWorld {
         String MyString = "Hello";
         System.out.print("\nMyString");
 
-        long myNumber = 123;
+        float myNumber = 123;
         System.out.println(myNumber);
+
+//        int x = 5;
+//        System.out.println(x++);
+//        System.out.println(x);
+
+//        int x = 5;
+//        System.out.println(++x);
+//        System.out.println(x);
+
+//        int class = 5
+
+//        String theNumberThree = "three";
+//        Object o = theNumberThree;
+//        int three = (int) o;
+
+//        int three = (int) "three";
+
+//        int x = 4;
+//        x = x + 5;
+//        x += 5;
+        //System.out.println(x);
+
+//        int x = 3;
+//        int y = 4;
+//        y = y * x;
+//        x *= y;
+
+//        int x = 10;
+//        int y = 2;
+//        x /= y;
+//        y -= x;
+
+
 
     }
